+++ comics/settings/base.py
@@ -72,7 +72,7 @@ INSTALLED_APPS = (
     'django.contrib.sites',
     'django.contrib.staticfiles',
     'django.contrib.webdesign',
-    'bootstrap',
+    'bootstrapform',
     'compressor',
     'registration',
     'south',